requirements.txt
streamlit
cohere
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
pdfplumber